Move random seed to top level constants file
