+++ src/czbenchmarks/constants.py
@@ -15,6 +15,8 @@ MODEL_WEIGHTS_CACHE_PATH = os.environ.get(
     "MODEL_WEIGHTS_CACHE_PATH", "~/.cz-benchmarks/weights"
 )
 
+# Constants
+RANDOM_SEED = 42
 
 # Derived constants
 def get_numbered_path(base_path: str, index: int) -> str:
