+++ src/czbenchmarks/tasks/constants.py
@@ -1,6 +1,5 @@
 # FIXME: most  should be set directly in metric function call
 #        and random seed should be set for entire library
-RANDOM_SEED = 42
 
 N_FOLDS = 5
