+++ src/czbenchmarks/tasks/label_prediction.py
@@ -22,7 +22,8 @@ from ..metrics import metrics_registry
 from ..metrics.types import MetricResult, MetricType
 from .base import BaseTask
 from .utils import filter_minimum_class
-from .constants import RANDOM_SEED, N_FOLDS, MIN_CLASS_SIZE
+from .constants import N_FOLDS, MIN_CLASS_SIZE
+from ...constants import RANDOM_SEED
 
 logger = logging.getLogger(__name__)
